Refactor init.pp to take the payloads and a schedule as params
Prior to this commit, all of the cron jobs were hardcoded and
not configurable.

After this commit, all of the cron jobs are configurable via a
single parameter that takes payloads and a schedule to run them
on.  This commit does not change the resulting cron jobs it just
moves the defaults from hardcoded in the manifest to parameters.