Allow running two GC types at the same time
Prior to this commit, you could not easily specify that you wanted
to run two types of GC in the same cron.  For example to run
expire_nodes and purge_nodes you'd have to set the payload to
'expire_nodes"],["purge_nodes' because the template added the
leading and trailing double quotes and brackets.

After this commit, we allow a string or Array[String] for
api_payload and then format the string so it looks like a json
array so we no longer need the leading and trailing quotes and
brackets as they are in the parameter when passed in.