Add a vacuum to the end of the purge_reports gc call
Prior to this commit, we realized that whenever you delete reports
you've created enough garbage in postgresql to warrant a vacuum,
however, with autovacuum you wait until some unknown amount of time
to perform that vacuum and in the meantime some of the data you just
read off of disk will be flushed by out to disk and then need to be
re-read.

After this commit, immediately after deleting reports we ask to vacuum
the reports table which should result in smaller more consistent
vacuums and they won't need to read the data from disk because it was
just read into memory when the delete ran.  So overall we expect a
small perf improvement and more consistent load on the server.