Create Puppetdb_gc::Cron_schedule to validate inputs
Prior to this commit, you could provide an arbitrary hash to the
gc_cron defined type and it would try to expand it onto the cron
resource.

After this commit, we have a new custom data type that validates
the input of the schedule as only the fields we want and of the
data types we want.