Disable vacuuming on the reports table by default
Prior to this commit, we were vacuuming the reports table after
running GC on reports by default.  However,  if you decided that
running report GC every 5 minutes was the most efficient thing to
do then running a VACUUM after that every 5 minutes probably
wouldn't be very efficient.

After this commit, we default to not vacuuming the report table
after reports GC is completed.